added friend search as a component
