+++ client/src/components/FriendList/index.js
@@ -1,20 +1,30 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import { useMutation } from '@apollo/client';
+import { REMOVE_FRIEND} from '../../utils/mutations';
+import FriendSearch from '../FriendSearch';
 
 
-const FriendList = ({ friendCount, username, friends }) => {
+const FriendList = ({ friendCount, username, friends, handleRemoveFriend }) => {
+  const [removeFriend] = useMutation(REMOVE_FRIEND);
+ 
   if (!friends || !friends.length) {
     return <p className="bg-dark text-light p-3">{username}, make some friends!</p>;
   }
 
+  const handleClickRemove = async (username) => {
+    handleRemoveFriend(username)
+  };
+
   return (
     <div>
-      
       {friends.map(friend => (
         <button className="btn w-100 display-block mb-2" key={friend._id}>
-          <Link to={`/profile/${friend.username}`}>{friend.username}</Link>
-        </button>
+          <Link to={`/profile/${friend.username}`}>{friend.username} </Link>  
+          <i className="fas fa-minus-circle fa-xs" onClick={() => handleClickRemove(friend.username)} ></i>
+        </button>      
       ))}
+      <FriendSearch />
     </div>
   );
 };
